flight_status/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
cachecontrol==0.14.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
confluent-kafka==2.5.0
cryptography==43.0.0
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
httplib2==0.22.0
idna==3.7
kafka-python==2.0.2
kombu==5.3.7
msgpack==1.0.8
multidict==6.0.5
pika==1.3.2
prompt-toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
redis==5.0.7
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
twilio==9.2.3
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
